Support init-depend clauses in unit contracts
This is a rather large and somewhat non-backwards-compatible change to unit
contracts. Unit contracts must now specify at least as many initialization
dependencies as the unit value being contracted, but may include more. These
new dependencies are now actually specified in the unit wrapper so that
they will be checked by compound-unit expressions.

Additionally, this commit adds more information to the first-order-check
error messages. If a unit imports tagged signatures, previously the errror
message did not specify which tag was missing from the unit contract. Now
the tag is printed along with the signature name.