Add support for space-efficient arrow contracts.
When a second contract is applied to a function, switch representation
for the wrapper and try eliding redundant checks. The resulting function
keeps a constant number of chaperone/impersonator wrappers regardless of
the number of contracts applied to it, and won't run any (provably)
redundant checks.

This avoids pathological cases where, e.g., a function crosses a
boundary inside a loop, and gets wrapped N times (or worse, 2^N).

The optimization currently only applies for fixed-arity functions and
contracts, and only for functions with known result-arity of 1. These
limitations are not fundamental.

Checking specific checks is not as optimized as for regular arrow
contracts yet. (Specifically: arity-specific wrappers and
tail-marks-match support is missing.) Again, not a fundamental
limitation.

In collaboration with Christophe Scholliers and Robby Findler.