Bail out of space-efficient mode when there are non-flat leaves in the supported trees of contracts
When there are higher-order contracts that don't support space-efficient mode at the leaves, it is possible
that merging contracts will drop contracts with different blame, thus leading to incorrect blame later
on when checking negative positions. The current fix for this is to only allow flat contracts at the
leaves of the space-efficient structure.

There are a number of failing tests for the above reasons, but it isn't entirely clear whether those
test cases are still meaningful.