Don't attach the space-efficient impersonator property unless necessary
This is probably a tradeoff, it might be faster to not attach the property, but
then checking if a value supports space-efficient mode may become slightly slower.
I suspect the cost of attaching the property is worse than checking for s-e support,
though, so this should be ok. Need measurements to be sure.