Change the collapsing strategy
When collapsing N contracts, split the contracts into their positive,
negative, and first-order componenets and merge each separately.
This allows us to detect when we can obviously skip parts of the
contracts to be merged. On the vector/c micro-benchmarks that add
the same contract 11 times, this change results in an almost 3x
improvement, although the collapsing is still MUCH slower
than attaching an 11th contract in regular racket.