Fix bugs in the collapsing strategy with higher order contracts
I'm less convinced that this is generally an optimization, but since the
lists of contracts that are collapsed have length at most 10 (or some other small number)
it's possible that we could be more aggressive to handle the situations that show up in
the benchmarks. Or maybe just revert to the original strategy.